Refactor initiator package (#59)
* refactor: (crypto) extracted code from `crypto.go` into `beacon.go` and `rsa.go`

* refactor initiator package

* finish operators map to slice refactor

* Refactor code to use multiline append statements in initiator and operator tests

* more refactors
