+++ pkgs/initiator/initiator_test.go
@@ -3,6 +3,7 @@ package initiator_test
 import (
 	"crypto/rsa"
 	"encoding/hex"
+	"encoding/json"
 	"errors"
 	"math/big"
 	"testing"
@@ -51,16 +52,19 @@ func TestStartDKG(t *testing.T) {
 	err := logging.SetGlobalLogger("debug", "capital", "console", nil)
 	require.NoError(t, err)
 	logger := zap.L().Named("operator-tests")
-	ops := make(map[uint64]initiator.Operator)
+	ops := initiator.Operators{}
 	version := "v1.0.2"
 	srv1 := test_utils.CreateTestOperatorFromFile(t, 1, examplePath, version)
 	srv2 := test_utils.CreateTestOperatorFromFile(t, 2, examplePath, version)
 	srv3 := test_utils.CreateTestOperatorFromFile(t, 3, examplePath, version)
 	srv4 := test_utils.CreateTestOperatorFromFile(t, 4, examplePath, version)
-	ops[1] = initiator.Operator{srv1.HttpSrv.URL, 1, &srv1.PrivKey.PublicKey}
-	ops[2] = initiator.Operator{srv2.HttpSrv.URL, 2, &srv2.PrivKey.PublicKey}
-	ops[3] = initiator.Operator{srv3.HttpSrv.URL, 3, &srv3.PrivKey.PublicKey}
-	ops[4] = initiator.Operator{srv4.HttpSrv.URL, 4, &srv4.PrivKey.PublicKey}
+	ops = append(
+		ops,
+		initiator.Operator{srv1.HttpSrv.URL, 1, &srv1.PrivKey.PublicKey},
+		initiator.Operator{srv2.HttpSrv.URL, 2, &srv2.PrivKey.PublicKey},
+		initiator.Operator{srv3.HttpSrv.URL, 3, &srv3.PrivKey.PublicKey},
+		initiator.Operator{srv4.HttpSrv.URL, 4, &srv4.PrivKey.PublicKey},
+	)
 	_, pv, err := rsaencryption.GenerateKeys()
 	require.NoError(t, err)
 	priv, err := rsaencryption.ConvertPemToPrivateKey(string(pv))
@@ -104,49 +108,51 @@ func TestStartDKG(t *testing.T) {
 
 func TestLoadOperators(t *testing.T) {
 	t.Run("test load happy flow", func(t *testing.T) {
-		ops, err := initiator.LoadOperatorsJson(jsonStr)
+		var ops initiator.Operators
+		err := json.Unmarshal(jsonStr, &ops)
 		require.NoError(t, err)
 		require.Len(t, ops, 4)
-		require.Equal(t, ops[4].Addr, "http://localhost:3033", "addr not equal")
+		require.Equal(t, ops[3].Addr, "http://localhost:3033", "addr not equal")
 		key3, err := crypto.ParseRSAPublicKey([]byte("LS0tLS1CRUdJTiBSU0EgUFVCTElDIEtFWS0tLS0tCk1JSUJJakFOQmdrcWhraUc5dzBCQVFFRkFBT0NBUThBTUlJQkNnS0NBUUVBdlFhZlo0ODJQYXRsYnRrOVdIb2MKZDBWdWNWWDk4QUlzenAvazlFTlYyQU82SVhQUXVqU1BtdUZrQTlibThsSllnWTJPb0lQU0RmK1JHWGNMc2R0VApzdEJhQ2JPL0pMOFlSejk4NURKejhBRlhDU0J3bW5mbzROSFptUjJGMVdMTE5CS2wzdVQ5Q1VLbC9RUnpKRFF1CjNNYVJ6eE5FVmdONWtvU1Nid0NxVDNDSCtjam5QU0pIeGhiaTNTaldOSnJFb3ZRUmN3ZUlpYXRrZEdVNWJOUkoKUW1LVldhYzhzVklYN2NDNE54V2RDNG1VM1RPK2Vlei90N2xVcnhSNjdnb21TbGdwaU5weFJ1M2dFajRkSWpINwpsZDlTYW1ObEJPeHV5N0lFMEJpdm5nSUdIKzVwcXZVTXhoM0N5WkVtMjFHd3JTRFhqcVpwWG92OEUwQkQ5eGY4ClN3SURBUUFCCi0tLS0tRU5EIFJTQSBQVUJMSUMgS0VZLS0tLS0K"))
 		require.NoError(t, err)
-		require.True(t, ops[3].PubKey.Equal(key3), "pubkey not equal")
+		require.True(t, ops[2].PubKey.Equal(key3), "pubkey not equal")
 	})
 	t.Run("test wrong pub key encoding", func(t *testing.T) {
-		_, err := initiator.LoadOperatorsJson([]byte(`[
+		var ops initiator.Operators
+		err := json.Unmarshal([]byte(`[
       {
         "id": 1,
         "public_key": "LS1CRUdJTiBSU0EgUFVCTElDIEtFWS0tLS0tCk1JSUJJakFOQmdrcWhraUc5dzBCQVFFRkFBT0NBUThBTUlJQkNnS0NBUUVBdkFXRFppc1d4TUV5MGNwdjhoanAKQThDMWNYZ3VseHkyK0tDNldpWGo3NThuMjl4b1NsNHV1SjgwQ2NqQXJqbGQrWkNEWmxvSlhtMk51L0FFOFRaMgpQRW1UZFcxcGp5TmV1N2RDUWtGTHF3b3JGZ1AzVWdxczdQSEpqSE1mOUtTb1Y0eUxlbkxwYlR0L2tEczJ1Y1c3CnUrY3hvZFJ4d01RZHZiN29mT0FhbVhxR1haZ0NhNHNvdHZmSW9RS1dDaW9MczcvUkM3dHJrUGJONW4rbHQyZWEKd1J1SFRTTlNZcEdmbi9ud0FROHVDaW55SnNQV0Q0NUhldG9GekNKSlBnNjYzVzE1K1VsWU9tQVJCcWtaSVBISAp5V25ORjZTS2tRalI2MDJwQ3RXTkZRMi9wUVFqblJXbUkrU2FjMHhXRVQ3UUlsVmYxSGZ2NWRnWE9OT05hTTlFClN3SURBUUFCCi0tLS0tRU5EIFJTQSBQVUJMSUMgS0VZLS0tLS0K",
         "ip": "http://localhost:3030"
       }
-    ]`))
+    ]`), &ops)
 		require.ErrorContains(t, err, "decode PEM block")
 	})
 	t.Run("test wrong operator URL", func(t *testing.T) {
-		_, err := initiator.LoadOperatorsJson([]byte(`[
+		var ops initiator.Operators
+		err := json.Unmarshal([]byte(`[
       {
         "id": 1,
         "public_key": "LS0tLS1CRUdJTiBSU0EgUFVCTElDIEtFWS0tLS0tCk1JSUJJakFOQmdrcWhraUc5dzBCQVFFRkFBT0NBUThBTUlJQkNnS0NBUUVBdkFXRFppc1d4TUV5MGNwdjhoanAKQThDMWNYZ3VseHkyK0tDNldpWGo3NThuMjl4b1NsNHV1SjgwQ2NqQXJqbGQrWkNEWmxvSlhtMk51L0FFOFRaMgpQRW1UZFcxcGp5TmV1N2RDUWtGTHF3b3JGZ1AzVWdxczdQSEpqSE1mOUtTb1Y0eUxlbkxwYlR0L2tEczJ1Y1c3CnUrY3hvZFJ4d01RZHZiN29mT0FhbVhxR1haZ0NhNHNvdHZmSW9RS1dDaW9MczcvUkM3dHJrUGJONW4rbHQyZWEKd1J1SFRTTlNZcEdmbi9ud0FROHVDaW55SnNQV0Q0NUhldG9GekNKSlBnNjYzVzE1K1VsWU9tQVJCcWtaSVBISAp5V25ORjZTS2tRalI2MDJwQ3RXTkZRMi9wUVFqblJXbUkrU2FjMHhXRVQ3UUlsVmYxSGZ2NWRnWE9OT05hTTlFClN3SURBUUFCCi0tLS0tRU5EIFJTQSBQVUJMSUMgS0VZLS0tLS0K",
         "ip": "wrongURL"
       }
-    ]`))
+    ]`), &ops)
 		require.ErrorContains(t, err, "invalid operator URL")
 	})
 }
 
 func generateOperators(ids []uint64) initiator.Operators {
-	m := make(map[uint64]initiator.Operator)
+	m := make([]initiator.Operator, 0, len(ids))
 	for _, i := range ids {
-		m[i] = initiator.Operator{
+		m = append(m, initiator.Operator{
 			Addr: "",
 			ID:   i,
 			PubKey: &rsa.PublicKey{
 				N: big.NewInt(1),
 				E: 0,
 			},
-		}
+		})
 	}
-
 	return m
 }
 
@@ -301,11 +307,12 @@ func TestRemoveTrailSlash(t *testing.T) {
 	}
 	]`)
 
-	ops, err := initiator.LoadOperatorsJson(jsonStr)
+	var ops initiator.Operators
+	err := json.Unmarshal(jsonStr, &ops)
 
 	require.Nil(t, err)
-	require.Equal(t, "http://localhost:3030", ops[1].Addr)
-	require.Equal(t, "http://localhost:3031", ops[2].Addr)
+	require.Equal(t, "http://localhost:3030", ops[0].Addr)
+	require.Equal(t, "http://localhost:3031", ops[1].Addr)
 }
 
 func TestDepositDataSigningAndVerification(t *testing.T) {
